Rename test support extension to a more meaningful name
