--- integration-tests-support/core-main/runtime/src/main/java/org/apache/camel/quarkus/core/runtime/support/MyPair.java
@@ -1,32 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.camel.quarkus.core.runtime.support;
-
-public class MyPair {
-    public final String key;
-    public final String val;
-
-    public MyPair(String key, String val) {
-        this.key = key;
-        this.val = val;
-    }
-
-    public static MyPair fromString(String input) {
-        String[] items = input.split(":");
-        return new MyPair(items[0], items[1]);
-    }
-}
