+++ integration-tests-support/custom-log-component/deployment/src/main/java/org/apache/camel/quarkus/core/support/deployment/SupportBuildStep.java
