+++ integration-tests-support/custom-log-component/runtime/src/main/java/org/apache/camel/quarkus/core/support/SupportRecorder.java
