+++ integration-tests-support/custom-routes-collector/deployment/src/main/java/org/apache/camel/quarkus/core/runtime/support/deployment/Feature.java
