+++ integration-tests-support/custom-routes-collector/runtime/src/main/java/org/apache/camel/quarkus/core/runtime/support/SupportRoutesCollector.java
