+++ integration-tests/core-main-collector/pom.xml
@@ -42,7 +42,7 @@
     <dependencies>
         <dependency>
             <groupId>org.apache.camel.quarkus</groupId>
-            <artifactId>camel-quarkus-integration-test-support-core-main-collector-ext</artifactId>
+            <artifactId>camel-quarkus-integration-tests-support-custom-routes-collector</artifactId>
         </dependency>
         <dependency>
             <groupId>org.apache.camel.quarkus</groupId>
