+++ integration-tests/core-main/pom.xml
@@ -62,7 +62,11 @@
         </dependency>
         <dependency>
             <groupId>org.apache.camel.quarkus</groupId>
-            <artifactId>camel-quarkus-integration-test-support-core-main-ext</artifactId>
+            <artifactId>camel-quarkus-integration-tests-support-custom-type-converter</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.camel.quarkus</groupId>
+            <artifactId>camel-quarkus-integration-tests-support-custom-main-listener</artifactId>
         </dependency>
         <dependency>
             <groupId>org.apache.camel.quarkus</groupId>
